All validation Added

+++ Frontend/src/pages/Blog.jsx
@@ -10,12 +10,13 @@ const Blog = () => {
   const [showForm, setShowForm] = useState(false);
   const [editingBlog, setEditingBlog] = useState(null);
   const [isAdmin, setIsAdmin] = useState(false);
+  const [loading, setLoading] = useState(true);
 
   useEffect(() => {
     fetchBlogs();
 
     const admin = checkAdmin();
-      setIsAdmin(admin);
+    setIsAdmin(admin);
   }, []);
 
   const fetchBlogs = async () => {
@@ -23,17 +24,30 @@ const Blog = () => {
       const response = await API.get("/blogs/");
       setBlogs(response.data.blogs);
     } catch (error) {
-      console.error("Error fetching blogs:", error.message || "An error occurred.");
+      toast.error("Connect to Internet");
+      console.error(
+        "Error fetching blogs:",
+        error.message || "An error occurred."
+      );
+    } finally {
+      setLoading(false);
     }
   };
 
   const handelDelete = async (id) => {
     try {
-      const response = await API.delete(`/blogs/${id}`); // Dynamic id
-      toast.success(response.data.message);
-      fetchBlogs();
+      const confirmed = window.confirm(
+        "Are you sure you want to delete the blog ?"
+      );
+      if (confirmed) {
+        const response = await API.delete(`/blogs/${id}`);
+        toast.success(response?.data?.message);
+        fetchBlogs();
+      }
     } catch (error) {
-      toast.error(error?.response?.data?.message || "Failed to delete the blog.");
+      toast.error(
+        error?.response?.data?.message || "Failed to delete the blog."
+      );
     }
   };
 
@@ -53,16 +67,27 @@ const Blog = () => {
     fetchBlogs();
   };
 
+  // Loading state
+  if (loading) {
+    return (
+      <div className="flex justify-center items-center h-screen">
+        <div className="text-center">
+          <p className="text-lg text-green-600 font-semibold">Loading...</p>
+        </div>
+      </div>
+    );
+  }
+
   return (
     <div className="container mx-auto p-8">
       <h1 className="text-4xl font-bold mb-4">Blogs</h1>
       {isAdmin && (
-      <button
-        onClick={() => setShowForm(true)}
-        className="bg-green-500 text-white px-4 py-2 rounded mb-4"
-      >
-        Add New
-      </button>
+        <button
+          onClick={() => setShowForm(true)}
+          className="bg-green-500 text-white px-4 py-2 rounded mb-4"
+        >
+          Add New
+        </button>
       )}
 
       {showForm ? (
